+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This release comes with new features and tons of unit tests. To get the full exp
 -   `CMDBCategory::delete()`: Marks entry in a multi-value category for a specific object as deleted
 -   `CMDBCategory::purge()`: Purge entry in a multi-value category for a specific object
 -   `CMDBCategory::recycle()`: Restore entry in a multi-value category for a specific object to "normal" state
+-   `CMDBDialog::create()`: Reference parent entry by its title (string) or by its identifier (integer)
 -   `Idoit::getAddOns()`: Read information about installed add-ons
 -   `CheckMKTags`: Read host tags by one or more objects from category `C__CATG__CMK_TAG`
 -   `API::rawRequest()`: Perform a low level API request